New files
Update for the new season. I'm going to be storing the data in an XML
for now. Have the .rng set up, some example files, and the start of the
new scraper (to a site that has all the information I need for pitchers
except no-hitters and perfect games (including holds and quality
starts)). Probably going to need to change the title of the repo now
that I'm using thebaseballcube.com.